+++ docs/include/samplecode/SpotifyClient/GetArtistInfo.py
@@ -1,0 +1,25 @@
+from spotifywebapipython import *
+from spotifywebapipython.models import *
+
+try:
+
+    CLIENT_ID:str = 'your_client_id'
+    CLIENT_SECRET:str = 'your_client_secret'
+
+    # create new spotify client instance.
+    spotify:SpotifyClient = SpotifyClient()
+
+    # generate a spotify client credentials access token (no scope, public data use only).
+    spotify.SetAuthTokenClientCredentials(CLIENT_ID, CLIENT_SECRET)
+    print('\nAuth Token:\n Type="%s"\n Scope="%s"' % (spotify.AuthToken.AuthorizationType, str(spotify.AuthToken.Scope)))
+    print('\nUser:\n DisplayName="%s"\n EMail="%s"' % (spotify.UserProfile.DisplayName, spotify.UserProfile.EMail))
+
+    # get Spotify artist biography page information.
+    artistId:str = '6APm8EjxOHSYM5B4i3vT3q'   # MercyMe
+    print('\nGetting details for Artist "%s" ...\n' % artistId)
+    artistInfo:ArtistInfo = spotify.GetArtistInfo(artistId)
+    print(str(artistInfo))
+
+except Exception as ex:
+
+    print("** Exception: %s" % str(ex))