+++ docs/include/samplecode/SpotifyClient/RemovePlaylist.py
@@ -1,0 +1,35 @@
+from spotifywebapipython import *
+from spotifywebapipython.models import *
+
+try:
+
+    # this sample requires an authorization token, as it requires security scope to accesses user data.
+
+    CLIENT_ID:str = 'your_client_id'
+    SPOTIFY_SCOPES:list = \
+    [
+        'user-read-email',
+        'user-library-read',
+        'user-library-modify',
+        'playlist-modify-private',
+        'playlist-modify-public'
+    ]
+
+    # create new spotify client instance.
+    spotify:SpotifyClient = SpotifyClient()
+
+    # generate a spotify authorization code with PKCE access token (with scope, private and public data use).
+    spotify.SetAuthTokenAuthorizationCodePKCE(CLIENT_ID, SPOTIFY_SCOPES)
+    print('\nAuth Token:\n Type="%s"\n Scope="%s"' % (spotify.AuthToken.AuthorizationType, str(spotify.AuthToken.Scope)))
+    print('\nUser:\n DisplayName="%s"\n EMail="%s"' % (spotify.UserProfile.DisplayName, spotify.UserProfile.EMail))
+
+    # remove playlist (ie. unfollow it).
+    playlistId:str = '4yptcTKnXjCu3V92tVVafS'
+    print('\nRemoving playlist id "%s"' % playlistId)
+    spotify.RemovePlaylist(playlistId)
+
+    print('\nPlaylist removed successfully')
+
+except Exception as ex:
+
+    print("** Exception: %s" % str(ex))